requirements.txt
kivy
openai
pytesseract
Pillow
PyPDF2
kivmob
sqlite3
